requirements.txt
Flask==1.1.2
Flask-PyMongo==2.3.0
Flask-RESTful==0.3.8
flask-restplus==0.13.0
flask-restx==0.1.1
mysql-connector==2.2.9
pymongo==3.4.0
python-editor==1.0.4
Werkzeug==0.16.1